Prevent errors by checking for display files
Adds a check for the display output files before reading them to prevent errors. Also changed: the DISP_PATH is slightly more general, and when switching is stopped, the number of display outputs found is reported to the user. This would help the user know if a particular output isn't being found (eg. 4 physical connections on the card but only 3 are reported indicates a problem finding 1 output)