Check if drivers are in use
Added a check for if the drivers are in use before unloading them. Previously a program like Folding @ Home could keep the drivers active and prevent the remove function from completing, causing a black screen. Now if remove is called while a program like F@H is active, the script will print an error, stop trying the removal and restart the display manager.